fix(csa-resource): persist resolved extra-readable paths (#3074)
